fix: auditoría de seguridad y calidad — hardening general
- CotizarRequest: precioUsd y pesoKg cambian de double a BigDecimal para evitar
  pérdida de precisión en cálculos de importación
- CotizadorEngine: extrae doubleValue() al inicio de calcular() — aritmética
  interna sigue en double sin cambiar la lógica existente
- CotizacionService: elimina valueOf() redundantes al persistir BigDecimal
- SolicitudService.calcularSugerencia: envuelve primitivos en BigDecimal.valueOf()
  para construir CotizarRequest con la nueva firma
- CotizadorEngineTest: actualiza helper req() para pasar BigDecimal al constructor
- EmailService: agrega esc() para escapar HTML en todos los campos de usuario,
  previniendo XSS en emails generados
- ListingRepository: corrige JPQL de búsqueda — LOWER/CONCAT en ambas columnas
  para búsqueda case-insensitive correcta
- MarketplaceService.calcularPrecioArs: null guard — retorna BigDecimal.ZERO
  si precioUsd es null, evita NPE en cron de actualización de precios
- V7__check_constraints.sql: CHECK constraints en estados de solicitudes, ítems
  y pedidos; constraint de integridad para origen de pedido

+++ src/main/java/com/hornetimports/cotizador/CotizacionService.java
@@ -41,8 +41,8 @@ public class CotizacionService {
         c.setUserId(user != null ? user.getId() : null);
         c.setProductoUrl(request.productoUrl());
         c.setNombreProducto(request.nombreProducto());
-        c.setPrecioUsd(BigDecimal.valueOf(request.precioUsd()));
-        c.setPesoKg(BigDecimal.valueOf(request.pesoKg()));
+        c.setPrecioUsd(request.precioUsd());
+        c.setPesoKg(request.pesoKg());
         c.setCategoria(request.categoriaId());
         c.setCostoTotalArs(BigDecimal.valueOf(resultado.desglose().totalArs));
         c.setDesglose(resultado.desglose());
