+++ src/main/java/com/hornetimports/cotizador/dto/CotizarRequest.java
@@ -2,11 +2,13 @@ package com.hornetimports.cotizador.dto;
 
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 public record CotizarRequest(
         @NotBlank  String productoUrl,
         @NotBlank  String nombreProducto,
-        @NotNull @Positive double precioUsd,
-        @NotNull @Positive @DecimalMax("30.0") double pesoKg,
+        @NotNull @Positive BigDecimal precioUsd,
+        @NotNull @Positive @DecimalMax("30.0") BigDecimal pesoKg,
         @NotBlank  String categoriaId,
         @NotBlank @Pattern(regexp = "particular|mayorista") String tipo,
         @NotBlank @Pattern(regexp = "completo|forwarding")  String tipoServicio,
